+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 - added an optional "cost limit" to `TMMemoryCache`, including new properties and methods
 - calling `[TMDiskCache trimToDate:]` with `[NSDate distantPast]` will now clear the cache
+- calling `[TMDiskCache trimDiskToSize:]` will now remove files in order of access date
 - setting the byte limit on `TMDiskCache` to 0 will no longer clear the cache (0 means no limit)